requirements.txt
Django==3.1.7
django-crispy-forms==1.11.2
pandas_datareader==0.9.0
scikit-learn==0.24.1
<<<<<<< HEAD
matplotlib==3.1
Keras
=======
matplotlib==3.4.1
tensorflow==2.5.0rc
jupyter
numpy
scipy
matplotlib
scikit-learn
sphinx
six
matplotlib
Pillow
quandl
keras
tweepy
textblob
Flask
pytest
finnhub-python
>>>>>>> b0a0eca7193014d6c0d2b37983213dd1dd1c0957
